Added validation for while user remove from class

+++ api/src/main/java/org/ednovo/gooru/domain/service/ClassServiceImpl.java
@@ -6,8 +6,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import org.ednovo.gooru.application.util.TaxonomyUtil;
 import org.ednovo.gooru.application.util.GooruImageUtil;
+import org.ednovo.gooru.application.util.TaxonomyUtil;
 import org.ednovo.gooru.core.api.model.ActionResponseDTO;
 import org.ednovo.gooru.core.api.model.User;
 import org.ednovo.gooru.core.api.model.UserClass;
@@ -17,8 +17,8 @@ import org.ednovo.gooru.core.constant.ConstantProperties;
 import org.ednovo.gooru.core.constant.ParameterProperties;
 import org.ednovo.gooru.domain.service.setting.SettingService;
 import org.ednovo.gooru.infrastructure.persistence.hibernate.ClassRepository;
-import org.ednovo.gooru.infrastructure.persistence.hibernate.party.UserGroupRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
@@ -162,11 +162,14 @@ public class ClassServiceImpl extends BaseServiceImpl implements ClassService, C
 
 	@Override
 	@Transactional(readOnly = false, propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
-	public void deleteUserFromClass(final String classUid, final String userUid) {
-		this.getClassRepository().deleteUserFromClass(classUid, userUid);
+	public void deleteUserFromClass(final String classUid, final String userUid, User user) {
+		UserClass userclass = this.getClassRepository().getClassById(classUid);
+		rejectIfNull(userclass, GL0056, CLASS);
+		if (userclass.getUserUid().equals(user.getGooruUId()) || user.getGooruUId().equals(userUid)) {
+			this.getClassRepository().deleteUserFromClass(classUid, userUid);
+		} else {
+			throw new AccessDeniedException(generateErrorMessage("GL0089"));
+		}
 	}
 
-	
 }
-		
-
